+++ tests/test_uint8_64bit.c
@@ -11,7 +11,7 @@
 #include <limits.h>
 #include "grb2.h"
 
-int uint8(unsigned char *p);
+unsigned long int uint8(unsigned char *p);
 
 int
 main()
